requirement.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirement.in -o requirement.txt
anyio==4.7.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
beautifulsoup4==4.12.3
    # via python-fasthtml
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via uvicorn
fastcore==1.7.26
    # via
    #   fastlite
    #   python-fasthtml
    #   sqlite-minutils
fastlite==0.0.13
    # via python-fasthtml
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via python-fasthtml
idna==3.10
    # via
    #   anyio
    #   httpx
itsdangerous==2.2.0
    # via python-fasthtml
oauthlib==3.2.2
    # via python-fasthtml
packaging==24.2
    # via fastcore
python-dateutil==2.9.0.post0
    # via python-fasthtml
python-dotenv==1.0.1
    # via uvicorn
python-fasthtml==0.10.1
    # via -r requirement.in
python-multipart==0.0.19
    # via python-fasthtml
pyyaml==6.0.2
    # via uvicorn
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
sqlite-minutils==4.0.3
    # via fastlite
starlette==0.41.3
    # via python-fasthtml
typing-extensions==4.12.2
    # via anyio
uvicorn==0.32.1
    # via python-fasthtml
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.3
    # via uvicorn
websockets==14.1
    # via uvicorn
